Fix npm version/license drift, add sync-version script and CI check
- package.json: version 0.4.2-dev -> 0.5.1-dev, license MIT -> OpenAll-1.0
- bin/sync-version: syncs VERSION -> package.json + pyproject.toml
- Makefile: add sync-version and check-version targets, add to ci
- ci.yml: add check-consistency job
- 13 tests for sync-version script

+++ lib/concerns/nudge_coordinator.py
@@ -50,7 +50,7 @@ class NudgeCoordinator(Concern):
         if not rec or rec.consecutive_ineffective <= 1:
             return self.COOLDOWN
         backoff_exp = min(rec.consecutive_ineffective - 1, 3)
-        return self.COOLDOWN * min(2**backoff_exp, self.MAX_BACKOFF_MULTIPLIER)
+        return int(self.COOLDOWN * min(2**backoff_exp, self.MAX_BACKOFF_MULTIPLIER))
 
     def _can_nudge(self, name: str, now: int) -> bool:
         rec = self._records.get(name)
